Update index.html, normalize.css and add style.css
- Updated index.html to include link to the custom stylesheet, added id
attributes to some elements to better control its style in the css and
added the href link to the wikipedia entry
- Updated normalize.css to include a new reset property for blockquote
since it had margin added to it by default
- Added a custom stylesheet that stylizes the site of which the
highlight is making the image responsive. Also, the image's display
property is set to display block since FCC required it to be so in order
to pass the tests. However the image remains responsive even without it
(would require removal of the margin property as well)